Moved all MPI communication functions in their own source file Added position exchange function plus test Added ekin and temp reduce function
